+++ visualization/app/app.config.ts
@@ -23,6 +23,7 @@ import { UpdateVisibleTopLabelsEffect } from "app/codeCharta/state/effects/updat
 import { ResetColorRangeEffect } from "app/codeCharta/state/store/dynamicSettings/colorRange/resetColorRange.effect"
 import { BlacklistSearchPatternEffect } from "app/codeCharta/ui/ribbonBar/searchPanel/searchBar/blacklistSearchPattern.effect"
 import { UpdateShowLabelsEffect } from "./codeCharta/state/effects/updateShowLabels/updateShowLabels.effect"
+import { UpdateAmountOfEdgePreviewsEffect } from "./codeCharta/state/effects/amountOfEdgePreviews/updateAmountOfEdgePreviews.effect"
 
 export const appConfig: ApplicationConfig = {
     providers: [
@@ -34,6 +35,7 @@ export const appConfig: ApplicationConfig = {
         provideEffects([
             UnfocusNodesEffect,
             AddBlacklistItemsIfNotResultsInEmptyMapEffect,
+            UpdateAmountOfEdgePreviewsEffect,
             OpenNodeContextMenuEffect,
             BlacklistSearchPatternEffect,
             ResetColorRangeEffect,